store db table 변경

See merge request brandy.hwang/final_project_2!37

+++ green_plate_db.sql
@@ -25,8 +25,8 @@ create table t_product(
 create table t_store(
 	id integer auto_increment primary key,
     name varchar(100) not null,
-    latitude long,
-    longitude long,
+    latitude double,
+    longitude double,
     phone_num varchar(100) not null
 );
